added missing type cast

+++ src/triplet_vq_vae.py
@@ -135,7 +135,7 @@ class TripletVectorQuantizedVAE(VectorQuantizedVAE):
                     tf.float32
                 ),
             2),  
-            tf.shape(inputs)[0]
+            tf.cast(tf.shape(inputs)[0], tf.float32)
         )
 
         return distances
